cleanup, extract component
