+++ src/constants/medical-journal.ts
@@ -1,0 +1,116 @@
+export const CANMEDS_ROLES = [
+    "Medical Expert",
+    "Scholar",
+    "Communicator",
+    "Professional",
+    "Leader",
+    "Health Advocate",
+    "Collaborator",
+];
+
+export const LEARNING_OBJECTIVES = [
+    {
+        category: "Objectives of Clerkship Reflections",
+        objectives: [
+            "1. Promote Self-Awareness",
+            "   • Understand your emotions, biases, and values in patient care",
+            "   • Identify how personal beliefs influence your clinical behavior",
+            "",
+            "2. Enhance Clinical Reasoning",
+            "   • Reflect on diagnostic and therapeutic decisions",
+            "   • Analyze what went well or what could have been improved",
+            "",
+            "3. Support Professional Identity Formation",
+            "   • Explore how experiences are shaping you as a future doctor",
+            "   • Reflect on how you embody roles like communicator, collaborator, advocate",
+            "",
+            "4. Develop Empathy & Patient-Centered Thinking",
+            "   • Consider patient perspectives, values, and lived experience",
+            "   • Reflect on your communication, especially in difficult or emotional cases",
+            "",
+            "5. Integrate Feedback and Learning",
+            "   • Reflect on feedback received and how it informed your development",
+            "   • Identify learning goals or knowledge gaps for future improvement",
+        ],
+    },
+    {
+        category: "Techniques for Effective Reflection",
+        objectives: [
+            "1. Gibbs Reflective Cycle",
+            "   • Description – What happened?",
+            "   • Feelings – What were you thinking/feeling?",
+            "   • Evaluation – What was good/bad?",
+            "   • Analysis – Why did it happen that way?",
+            "   • Conclusion – What did you learn?",
+            "   • Action Plan – What will you do differently next time?",
+            "",
+            "2. The DIEP Model",
+            "   • Describe the experience",
+            "   • Interpret what it means",
+            "   • Evaluate its significance to your development",
+            "   • Plan how you'll apply the insight",
+            "",
+            "3. Free-Writing or Journaling",
+            "   • More informal but can be deeply introspective",
+            "   • Often useful right after emotional or challenging cases",
+            "",
+            "4. CanMEDS or ACGME Role-Based Reflection",
+            "   • Align your reflection with roles like 'Scholar', 'Communicator'",
+            "   • Some schools require or recommend mapping reflections to these roles",
+        ],
+    },
+    {
+        category: "Best Practices",
+        objectives: [
+            "• Focus on one meaningful encounter rather than a vague overview",
+            "• Be honest — it's about insight, not perfection",
+            "• Include both emotional and cognitive elements",
+            "• Tie your reflection to actionable learning or professional growth",
+            "• Avoid patient identifiers — maintain confidentiality",
+        ],
+    },
+];
+
+export const LEARNING_OBJECTIVES_DROPDOWN = [
+    "Appropriate consideration of broader differential diagnoses",
+    "Suggest rule out/ rule in investigations and management",
+    "Develop systematic approach to common presentations",
+    "Recognize critical vs non-critical patients",
+    "Perform focused history and physical examination",
+    "Document patient encounters accurately",
+    "Develop and implement management plans",
+    "Follow-up on investigation results",
+    "Clear communication with patients and families",
+    "Effective handover to other healthcare providers",
+    "Appropriate consultation with specialists",
+    "Documentation of clinical reasoning",
+];
+
+//todo: move to a db table somewhere
+export const ROTATIONS = [
+    "Emergency Medicine",
+    "Internal Medicine - MTU",
+    "Internal Medicine - Selective",
+    "General Surgery",
+    "Pediatrics",
+    "Obstetrics/Gynecology",
+    "Psychiatry",
+    "Family Medicine - Rural",
+    "Family Medicine - Urban",
+    "Vascular Surgery",
+    "Plastic Surgery",
+    "Anasthesia",
+    "NICU",
+    "Elective",
+    "Neurology",
+];
+
+export const LOCATIONS = [
+    "Emergency Department",
+    "Inpatient Ward",
+    "Outpatient Clinic",
+    "Operating Room",
+    "Intensive Care Unit",
+    "Other",
+    "Admissions",
+];
